+++ CHANGELOG.md
@@ -16,6 +16,7 @@ with the `YY.MINOR.MICRO` scheme.
 ### SQL
 - SQL supports `ORDER BY ... NULLS FIRST/LAST` operation.
 - SQL supports `INSERT INTO ... ON CONFLICT DO FAIL/REPLACE/NOTHING` operation for globally distributed tables.
+- SQL supports modulo operator `%` for integers.
 
 ### Pgproto